tasks 8 up3

+++ 0x00-MySQL_Advanced/7-average_score.sql
@@ -8,7 +8,7 @@ CREATE PROCEDURE ComputeAverageScoreForUser(
     IN user_id INT
 )
 BEGIN
-    DECLARE avg_score FLOAT;
+    DECLARE avg_score INT;
     SELECT AVG(score) INTO avg_score
     FROM corrections
     WHERE user_id = user_id;
